Merged PR 19230: engineering: Refactor API primitives for easier reuse
# 🔍 Description

- No external API changes.
- Refactor storage sizes into a standalone type (`ByteCount`.
- Proper implementation of JsonSchema for the types.
- Docbuilder improvements to handle new type.

# 🤔 Rationale

- Allows for reuse.
- Better documentation & doc generation.
- Create a module for other possible primitives (eg, time durations, checksums)

----
#### AI description  (iteration 1)
#### PR Classification
This pull request involves a significant refactor of the API primitives to enhance their reusability.

#### PR Summary
The changes in this pull request primarily focus on the introduction of a new `ByteCount` struct in the `bytes.rs` file, which provides a more reusable and efficient way to handle byte count operations. The `PartitionSize` struct in the `partitions.rs` file and the `sample_host_configuration` function in the `sample_hc.rs` file have been updated to use this new struct.
- Introduced a new `ByteCount` struct in `bytes.rs` that includes methods for conversion, human-readable formatting, and serialization/deserialization.
- Removed redundant byte count handling code from `partitions.rs` and replaced it with the new `ByteCount` struct.
- Updated `sample_host_configuration` function in `sample_hc.rs` to use the new `ByteCount` struct for partition size specification.