Merged PR 19287: engineering: log explainer script for trident's full log
# 🔍 Description

While looking at a full log from a customer I remembered I meant to create a tool like this.
It renders the JSON stream into more legible output:
![image.png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/19287/attachments/image.png)

When inside vscode, the filenames function as a link and take you directly to the logging call in the source code. :)

I also noticed the logging for binaries is... _subpar_ so I changed it a bit to make it easier to read.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature for log explanation and improves the logging system in the Trident project.

#### PR Summary
The pull request adds a Python script for parsing and colorizing Rust logs, and modifies the logging format in the `osutils` module.
- In `/scripts/log-explainer`, a new Python script is added which parses and colorizes Rust logs, improving readability and debugging efficiency.
- In `/osutils/src/exe.rs`, the logging format for command execution is updated to include a more detailed report, enhancing the traceability of command execution.

Related work items: #7922