Merged PR 19310: Bug: Fixing name to select the correct iso
# 🔍 Description

The `run-netlaunch-container` rule in the makefile was using the wrong ISO.

# 🤔 Rationale

For `run-netlaunch-container`  we want to check `trident-containerhost-mos.iso`, as stated in the dependencies, instead of `trident-mos.iso`

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a bug fix aimed at resolving an issue with the datastore failing for trident when running from a container.

#### PR Summary
The pull request modifies the Makefile to correct the iso name used when running the netlaunch container. This change is expected to fix the datastore creation failure when trident runs from a container.
- In the Makefile, the iso name in the `run-netlaunch-container` target has been changed from `bin/trident-mos.iso` to `bin/trident-containerhost-mos.iso`.

Related work items: #7946