Merged PR 19294: engineering: Point Trident to a custom host config on the host for A/B update testing
# 🔍 Description

This PR modifies the A/B update testing logic, to create a custom host config on the host for A/B update testing.

# 🤔 Rationale

With the incoming SELinux changes, there will no longer be a RW /etc on the host. So, the A/B update testing logic needs to be updated, to point Trident to an alternative host config file, when re-running the agent.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature for A/B update testing by creating a custom host config on the host.

#### PR Summary
The changes in this pull request allow for the creation of a custom host config on the host for A/B update testing. This is achieved by modifying the Trident config and triggering an A/B update.
- In `e2e_tests/helpers/ab_update.py`, the `update_host_trident_config` function was modified to update the host's Trident config via a `sed` command. The `trigger_ab_update` function was also updated to construct the full path to the host's Trident config file and update it.
- In `.pipelines/templates/stages/deployment_testing/common_tasks/e2e-test-run.yml`, steps were added to copy the Trident host config from `/etc/trident/config.yaml` to a new location. The `trigger_ab_update` function was then called with the new host config path.
- The `trident_run` function in `e2e_tests/helpers/ab_update.py` was updated to provide the `-c` arg, the full path to the RW host config.
- The `main` function in `e2e_tests/helpers/ab_update.py` was updated to accept the new `destination_dir` and `host_config` arguments.
- The `HOST_TRIDENT_CONFIG_PATH` constant was removed from `e2e_tests/helpers/ab_update.py` as it is no longer needed.

Related work items: #7941