Merged PR 19209: feature: Adding optional wait for RAID array syncing
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?

Adding an optional field to RAID config `syncTimeout` which makes Trident poll each raid device's `md/sync_action` file to check when it is done from `resync` to `idle`. We fail if the raid devices are not finished syncing within the provided timeout in the host config. The timeout value is in minutes.

Here is an example of what the sync_action file's output, which is handled by `mdadm` looks like as RAID devices are syncing:
![image (2).png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/19209/attachments/image%20%282%29.png)

# 🤔 Rationale

Why is this PR needed?

Request for optional wait for RAID syncing.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

- Add tests
- update docs

- #0000

# 🗒️ Notes

Adding optional wait for RAID array syncing

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature to the software RAID configuration.

#### PR Summary
The changes in this pull request add an optional wait for RAID array syncing, improving the reliability of the RAID setup process. The timeout for the sync operation can be configured, allowing for flexibility based on the specific needs of the RAID array.
- In `/src/modules/storage/raid.rs`, two new functions `wait_for_raid_sync` and `get_raid_devices` were added to handle the RAID sync operation and retrieval of RAID devices respectively.
- In `/src/modules/storage/encryption.rs`, the RAID configuration was updated to include the `sync_timeout` option.
- In `/trident_api/src/config/host/storage/raid.rs`, a new field `sync_timeout` was added to the `Raid` struct to allow for the configuration of the sync timeout.
- In `/trident_api/src/samples/sample_hc.rs`, the sample host configurations were updated to include the `sync_timeout` option.
- In `/src/modules/storage/mod.rs` and `/trident_api/src/config/host/storage/mod.rs`, the test configurations were updated to include the `sync_timeout` option.

Related work items: #6792