Merged PR 19320: engineering: Refactor the logic to copy runtime OS images in e2e-test-run.yml into a script
# 🔍 Description

This PR refactors the logic to copy runtime OS images into a separate script for more clean YAML.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a code refactor aimed at improving the readability and maintainability of the codebase.

#### PR Summary
The pull request refactors the logic for copying runtime OS images in the `e2e-test-run.yml` file into a separate script. This change addresses the linked work item, which suggested creating a temporary `destinationDirectory` in the testing environment, adding all necessary files there, and then performing a single `scp` operation to transfer the entire directory.
- In `copy-runtime-os-images.sh`, a new script was created to handle the copying of runtime OS images based on the `verityRequired` parameter.
- In `e2e-test-run.yml`, the logic for copying runtime OS images was replaced with a call to the new script `copy-runtime-os-images.sh`.
- The variable `verityExists` was renamed to `verityRequired` for better clarity.

Related work items: #7943