Merged PR 19333: engineering: Add logs when staging of clean install or update succeeded
# 🔍 Description

This PR adds two logs, to inform the user that the staging of clean install or update succeeded. This is especially relevant to the scenario where only staging of update is requested, since currently, there are no logs confirming that the provisioning succeeded.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is primarily focused on enhancing the logging mechanism in the application.

#### PR Summary
The changes in this pull request add logging statements to indicate the successful staging of clean installs and updates. This will improve the traceability and debugging process.
- In the file `/src/modules/mod.rs`, a log statement has been added after the staging of a clean install to indicate its success.
- Similarly, another log statement has been added in the same file to indicate the successful staging of an update.

Related work items: #7953