Merged PR 19338: engineering: Improve log-explorer
# 🔍 Description

- Add regex-based search, with grep-like highlighting. (Match is done per log, not per line!)
- Add capability to connect to an ssh server to retrieve the log.

Example:
![image.png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/19338/attachments/image.png)

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature to improve the log-explorer script.

#### PR Summary
The pull request enhances the log-explorer script by adding new functionalities such as the ability to parse and colorize Rust logs, and the ability to search for specific logs using regex. It also allows for the extraction of log files from a remote host.
- In `/scripts/log-explorer`, a new script is added that includes the ability to parse and colorize Rust logs, search for specific logs, and extract log files from a remote host.
- The existing script in `/scripts/log-explainer` is removed, as its functionalities are now included in the improved log-explorer script.

Related work items: #7922