Merged PR 19335: engineering: Don't publish docker image
# 🔍 Description

We aren't using the docker image for anything. This change retains the build step to make sure it keeps working but skips publishing it anywhere. That should eliminate the need to pass in credentials to publish to the ACR

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a code cleanup operation, specifically removing the functionality to publish Docker images from the pipeline.

#### PR Summary
The changes in this pull request remove the steps in the pipeline that were responsible for publishing Docker images. This simplifies the pipeline and reduces potential security risks associated with publishing images.
- In the file `/.pipelines/templates/stages/trident_rpms/release.yml`, the code for logging into the Docker registry, tagging the Docker image, and pushing the image to the registry has been removed.
- The display name of the pipeline step has been updated to reflect the new functionality: it now only builds the Docker image, and no longer publishes it.

Related work items: #7955