Merged PR 19349: engineering: Removing AUTO option for buildPurpose parameter in pipelines
# 🔍 Description

Remove ability to automatically define the build purpose (reason to run the E2E template) in the trident pipelines as the option is no longer used. Removing this allows to use the parameter in compile time operations.

# 🤔 Rationale

The `buildPurpose` was originally introduced to define the stages needed to be run on the e2e template, which have been replaced with `targetStage`. The other purpose that currently provides is define the level of testing that we desire, but now this is defined from each of the individual pipelines. We no longer use or need the `buildPurpose` to be automatically defined by the reason the pipeline run was triggered, which was defined at runtime. By removing this, we now can use `buildPurpose` for compile time operations and avoid running in difficult to debug situations.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a code cleanup that removes the "AUTO" option for the `buildPurpose` parameter in various pipeline templates.

#### PR Summary
The changes in this pull request simplify the build process by removing the automatic determination of the `buildPurpose` parameter. The default value for `buildPurpose` has been changed to "post_merge" in most cases.
- In `/.pipelines/templates/stages/deployment_testing/common_tasks/get-tests.yml`, the script for automatically setting the `buildPurpose` based on the build reason has been removed. The `buildPurpose` is now directly taken from the parameters.
- The "AUTO" option for `buildPurpose` has been removed from the parameters in `/.pipelines/templates/stages/deployment_testing/common_tasks/e2e-test-run.yml`, `/.pipelines/templates/stages/deployment_testing/vm/netlaunch-testing.yml`, and `/.pipelines/templates/stages/deployment_testing/baremetal/baremetal-testing.yml`.
- In `/.pipelines/templates/stages/deployment_testing/vm/netlaunch-testing.yml` and `/.pipelines/templates/stages/deployment_testing/baremetal/baremetal-testing.yml`, the `buildPurpose` variable is now directly taken from the parameters instead of being set by a previous task.

Related work items: #7793