Merged PR 19168: engineering: Add staging/finalizing A/B update flow to E2E testing
# 🔍 Description

This PR introduces a testing flow where A/B update is staged and finalized separately.

# 🤔 Rationale

This testing flow is needed because it may expose a different set of regressions than the standard flow, where A/B update is staged and finalized on a single Trident run.

Added structure for future testing

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature to the E2E testing process by adding a staging/finalizing A/B update flow.

#### PR Summary
The pull request enhances the E2E testing process by implementing a staged A/B update flow, which includes copying runtime OS images, triggering and validating the staging of A/B updates, and finalizing the A/B updates. It also introduces a new parameter to specify the stage type in the testing process.
- In `e2e-test-run.yml`, the code for performing A/B updates has been significantly expanded, including steps for copying runtime OS images, triggering and validating the staging of A/B updates, and finalizing the A/B updates.
- In `baremetal-testing.yml` and `netlaunch-testing.yml`, a new parameter `stageType` has been introduced to specify the stage type in the testing process.
- In `e2e-template.yml`, the `stageType` parameter is passed to the `netlaunch-testing.yml` and `baremetal-testing.yml` templates.

Related work items: #7793