Merged PR 19364: engineering: Check licenses in CI
# 🔍 Description

This PR makes sure that we only use dependencies with a select list of licenses.

It is OK to add to the list, but the idea is that we won't accidentally include a dependency that has an unsuitable license. The current list is:

* MIT
* Apache-2.0
* BSD-3-Clause
* ISC
* Zlib
* Unicode-DFS-2016

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature to validate licensing usage by all Open Source Software (OSS) dependencies.

#### PR Summary
The pull request adds a new configuration file and modifies the CI pipeline to check for license compliance of all dependencies. It also marks certain crates as non-publishable.
- Added a new configuration file `deny.toml` which specifies the allowed licenses and sets up rules for license checking.
- Modified the CI pipeline in `check.yml` to install `cargo-deny` and check licenses of all dependencies.
- Marked crates in `docbuilder/Cargo.toml`, `osutils/Cargo.toml`, `pytest/Cargo.toml`, `pytest_gen/Cargo.toml`, `setsail/Cargo.toml`, and `trident_api/Cargo.toml` as non-publishable by adding `publish = false`.

Related work items: #6220