Merged PR 19406: engineering: Update to changes in test-images
# 🔍 Description

This PR aligns with the changes on https://dev.azure.com/mariner-org/ECF/_git/test-images/pullrequest/19405.

# 🤔 Rationale

We want to use the download templates, so it is easier to select base image and MIC versions.

# 📝 Checks

- Running trident-pr-e2e pipeline.

TODO:

- Update MIC download script used by the template to be able to use release option and set that as default in this test-images and trident pipelines.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is primarily about updating the test image build process in the Trident pipelines.

#### PR Summary
The changes in this pull request aim to modify the way test images are built and managed in the Trident pipelines. The modifications are mainly in the `.pipelines/templates/stages/build_image/trident-testimg.yml` file.
- In `/.pipelines/templates/stages/build_image/trident-testimg.yml`, the `targetPath` for downloading Trident RPMs has been updated, and the parameters for the `trident-testimg-template.yml` template have been significantly altered. This includes changes to the `outputDirectory`, `micBuildType`, `micVersion`, `baseimgBuildType`, `baseimgVersion`, and `baseimgAzureLinuxVersion`.
- The variables in the same file have also been updated, with changes to `ob_outputDirectory`, `micBuildType`, `micVersion`, `baseimgBuildType`, `baseimgVersion`, and `baseimgAzureLinuxVersion`. The changes reflect a shift towards using the pipeline workspace for output and specifying versions and build types for the MIC and base image.

Related work items: #7986