Merged PR 19355: feature: Configure users in MOS
# 🔍 Description

Allow for user configuration in MOS.

# 🤔 Rationale

Mostly to be able to add SSH keys to the MOS.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature that allows users to be configured in the Management Operating System (MOS).

#### PR Summary
The changes in this pull request primarily involve the addition of new code blocks to handle user configuration in the MOS, as well as updates to the documentation and schema to reflect these changes.
- In `/src/modules/osconfig/mod.rs`, a new `MosConfigModule` struct and associated methods were added to handle user configuration in the MOS.
- The `/docs/Reference/Host-Configuration/API-Reference/ManagementOs.md` file was updated to include the new `users` field in the management OS configuration.
- The `/trident_api/schemas/host-config-schema.json` and `/trident_api/src/config/host/os/mod.rs` files were updated to include the new `users` field in the schema and configuration respectively.
- In `/src/modules/osconfig/users.rs`, the `ssh_global_config` function was updated to restart the sshd unit after writing the global user sshd config.
- The `/osutils/src/systemd.rs` file was updated to include a new function `restart_unit` to restart a systemd unit.

Related work items: #7961