Merged PR 19425: engineering: FT strict cleanup enforcement
# 🔍 Description

Make sure FTs are not leaving anything behind!

This extra check makes sure the disks are fully clean after going through wipefs and that the kernel doesn't still have references to them.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is focused on code cleanup and enforcing strict cleanup checks in functional tests.

#### PR Summary
The changes in this pull request enhance the cleanup process in functional tests by ensuring that the disk is clean after each test. This is achieved by adding a new function `assert_clean_disk` that checks the disk status.
- In `functional_tests/conftest.py`, the `wipe_sdb` function has been modified to call `assert_clean_disk` after wiping the disk.
- A new function `assert_clean_disk` has been added in `functional_tests/conftest.py` to check if the disk is clean after each test. It uses `lsblk` command to fetch disk information and asserts that there are no child partitions and the disk type is None.

Related work items: #7999