Merged PR 19175: engineering : Update selinux policy that works for e2e tests (non-verity)
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?
engineering: Updating the SELinux policy to enable support for different host configurations that customer can use.

# 🤔 Rationale
Why is this PR needed?
This Policy is focused on  end-to-end test configurations in the Trident project.

PR Summary

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)
Pipeline run with selinux in enforcing:
 https://dev.azure.com/mariner-org/ECF/_build/results?buildId=581741&view=results

# 📌 Follow-ups

TODO:

-  Add more filecon rules and update the policy for /var/lib folder access
- Fix AB update e2e test failure on ab-update with verity config seen because of selinux policy constraint violation.

# 🗒️ Notes

update policy

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is focused on updating the SELinux policy to enable support for end-to-end tests in the Trident project.

#### PR Summary
The pull request significantly expands the SELinux policy rules in the `/trident-selinuxpolicies.cil` file to ensure compatibility with end-to-end tests.
- Added numerous new permissions for various types of files and directories, enhancing the flexibility and functionality of the SELinux policy.
- Removed some redundant or unnecessary permissions, streamlining the policy.
- Modified several existing permissions to better align with the requirements of the end-to-end tests.
- These changes are crucial for re-enabling SELinux support in Trident, as outlined in the linked work item.

Related work items: #6775