Merged PR 19435: bug: fix clean-install datastore path
# 🔍 Description

Skip check for datastore path changes in clean install. This only make sense for updates.
Previously we checked against an option, since !19377 we check against the HC in the `spec` field of HS, which, on clean installs gets set to the default for `PathBuf`s (`""`), and the check always fails unless you use the default path.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a bug fix, addressing an issue with the clean-install datastore path in the management module.

#### PR Summary
The changes in this pull request ensure that the datastore path can only be changed during a clean install, preventing potential errors during other servicing types.
- In `src/modules/management.rs`, the `planned_servicing_type` parameter is now used to check if the servicing type is a clean install before allowing changes to the datastore path.
- If the servicing type is not a clean install, the current and new datastore paths are compared, and an error is returned if they differ.

Related work items: #7979