Merged PR 19457: bug: relax uuid parsing coming from block devices
# 🔍 Description

Replace `Uuid` with a new enum `OsUuid` which contains both a conforming UUID variant and a "relaxed" variant for unique identifiers that do not conform with proper UUIDs.

Previously when we encountered such strings we would fail completely. With this change, we should be able to parse anything and leave the decision of breaking or not to whoever is consuming the parsed data.

Related work items: #8021