Merged PR 19413: engineering: Modify logic for deploying images
# 🔍 Description

This PR updates the logic for deploying images during clean install and A/B update, in the following way:

1. An outdated object InternalImage is no longer being used in the logic. Trident now relies on the new objects in the host configuration: FileSystem and VerityFileSystem. To implement this, I incorporated the work that @<Paco Huelsz Prince> did in his feature branch, to implement image() and esp_image() for FileSystemSource, etc.: https://dev.azure.com/mariner-org/ECF/_git/trident?path=/&version=GBfrhuelsz/dualboot/init.
2. Trident now deploys all the images that are expected to be updated, during clean install and A/B update. In other words, if
an A/B update has been selected, Trident will assume that all the A/B update images, i.e. A/B volume pairs, in the host configuration have been updated.

# 🤔 Rationale

First, this PR is needed since we're shifting away from InternalImage to using FileSystem and VerityFileSystem, following the change in the API. Second, the assumption that all the images have been updated in the host config if A/B update is selected, is helpful for consistency and clear error logging. In the future, when the block device UUID check is implemented, to ensure that all the images are from the same build, this will allow to return a comprehensive and clear error to the user when they forget to update all relevant images in the host config.

# 📌 Follow-ups

TODO:

- Add unit tests for the remaining new functions:
1. test_get_esp_images()
2. test_get_updated_images()
3. test_get_images_from_filesystems()
4. test_get_ab_volume_pair_images()
5. test_get_images()

----
#### AI description  (iteration 1)
#### PR Classification
This pull request modifies the flow to deploy images during clean install and A/B update in the software development repository.

#### PR Summary
The pull request primarily modifies the image deployment process during clean install and A/B update. It also includes changes to the test cases and the filesystem configuration.
- Major modifications in `src/modules/storage/image/mod.rs` include the removal of the `get_undeployed_images` function and the addition of new functions `needs_ab_update`, `get_images_for_ab_update`, and `get_images_for_clean_install`. These changes aim to improve the image deployment process during clean install and A/B update.
- In `src/modules/storage/image/systemd_sysupdate.rs`, the `get_update_partition_id` function is updated to handle the device_id instead of the target_id. This change is part of the modifications to the image deployment process.
- The `trident_api/src/config/host/storage/filesystem.rs` file is updated to add new functions `image` and `esp_image` to the `FileSystemSource` struct. These functions return the image associated with the filesystem source.

Related work items: #7829, #7858