Merged PR 19339: bug: Handle sub-mounts
# 🔍 Description

We use `mount --rbind` instead of `mount --bind` to be able to handle nested mounts or sub-mounts.

# 🤔 Rationale
The `--rbind` option in the mount command creates a new mount point for a specific directory, but also includes all sub-mounts from the source directory (which `--bind` doesn't).

The `initialize_new_root` function, sets up the new root filesystem. The function calls bind_mount, which runs the Linux `mount` command. The switch from `--bind` to `--rbind` is needed when trident is running from a container, because we already [mounted some directories for the container to modify in the host](https://dev.azure.com/mariner-org/ECF/_git/trident?path=/trident-mos/files/download-trident-container.service&version=GBmain&line=9&lineEnd=9&lineStartColumn=1&lineEndColumn=243&lineStyle=plain&_a=contents).

# 📝 Checks

Manually checked that when trident is running on a container from the Provisioning OS the deployment succeeded:
 ![image.png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/19339/attachments/image.png)
![image (2).png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/19339/attachments/image%20%282%29.png)

# 📌 Follow-ups

TODO:
Reactivate Rerun test after:
- [Task 8244](https://dev.azure.com/mariner-org/ECF/_workitems/edit/8244/)
- [Task 8245](https://dev.azure.com/mariner-org/ECF/_workitems/edit/8245/)

# 🗒️ Notes
This PR is to unblock the use of trident from a container in the Provisioning OS.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a bug fix, addressing issues with the handling of sub-mounts in the datastore when Trident runs from a container.

#### PR Summary
The changes in this pull request improve error handling and logging in the datastore and modify the bind mount function to handle sub-mounts correctly.
- In `/src/datastore.rs`, error handling has been added to check if the new datastore path exists and return an error if it doesn't.
- In `/osutils/src/mount.rs`, the argument for the mount command has been changed from `--bind` to `--rbind` to correctly handle sub-mounts.
- In `/src/modules/mod.rs`, additional logging has been added to provide more information about the host status and the datastore path during the clean install process.

Related work items: #7946