Merged PR 19705: engineering: Update MIC and base image versions
# 🔍 Description

Updates the MIC and base image version numbers to match what test-images are using.

Add new watch-virtdeploy makefile target to help with monitoring VM boot progress.

# 🤔 Rationale

Ensure that Trident works with the latest versions.

----
#### AI description  (iteration 1)
#### PR Classification
Version update for base image and MIC package.

#### PR Summary
This pull request updates the versions of the base image and the MIC package in the Makefile.
- `Makefile`: Updated `BASE_IMAGE_VERSION` to `2.0.20240628`.
- `Makefile`: Updated `MIC_PACKAGE_VERSION` to `0.5.0-preview.611840`.

Related work items: #7949